@@ -34,6 +34,7 @@ import com.example.thesis.validations.BirthDayValidation;
 import com.example.thesis.validations.BrgyValidation;
 import com.example.thesis.validations.CityValidation;
 import com.example.thesis.validations.DateValidation;
+import com.example.thesis.validations.LastNameValidation;
 import com.example.thesis.validations.NameValidation;
 import com.example.thesis.validations.PhoneNumberValidation;
 import com.example.thesis.validations.StringValidation;
@@ -42,6 +43,7 @@ import com.example.thesis.validations.Validation;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
@@ -128,7 +130,8 @@ public class EditProfileActivity extends AuthenticatedActivity {
 
         // TODO: Add/Remove change password validation
         changePasswordButton.setOnClickListener(v -> {
-
+            Intent intent = new Intent(EditProfileActivity.this, ChangePassword.class);
+            startActivity(intent);
         });
     }
 
@@ -137,14 +140,22 @@ public class EditProfileActivity extends AuthenticatedActivity {
         currentUser.setFirstname(editTextFirstName.getText().toString());
         currentUser.setLastname(editTextLastName.getText().toString());
 
-        String bdayText = editBirthday.getText().toString();
-        if(!bdayText.isEmpty()) {
-            Instant instant = Instant.parse(bdayText + context.getResources().getString(R.string.instant_zero_time));
-            ZoneId zoneId = ZoneId.of(context.getResources().getString(R.string.used_time_zone));
-            ZonedDateTime zonedDateTime = instant.atZone(zoneId);
-            Calendar calendar = Calendar.getInstance();
-            calendar.set(zonedDateTime.getYear(), zonedDateTime.getMonthValue(), zonedDateTime.getDayOfMonth());
-            currentUser.setDateofbirth(calendar.getTime());
+        String birthdayText = editBirthday.getText().toString();
+        if(!birthdayText.isEmpty()) {
+            try {
+                Instant instant = Instant.parse(birthdayText + context.getResources().getString(R.string.instant_zero_time));
+                ZoneId zoneId = ZoneId.of(context.getResources().getString(R.string.used_time_zone));
+                ZonedDateTime zonedDateTime = instant.atZone(zoneId);
+                Calendar calendar = Calendar.getInstance();
+                calendar.set(zonedDateTime.getYear(), zonedDateTime.getMonthValue(), zonedDateTime.getDayOfMonth());
+                currentUser.setDateofbirth(calendar.getTime());
+            }
+            catch(DateTimeParseException dateTimeParseException){
+                message(getString(R.string.error_invalid_date));
+                revertSavedUserFields();
+                progressDialog.dismiss();
+                return;
+            }
         }
 
         currentUser.setPlaceofbirth(editBirthPlace.getText().toString());
@@ -158,6 +169,18 @@ public class EditProfileActivity extends AuthenticatedActivity {
         currentUser.setFathersname(editFathersName.getText().toString());
 
         saveUpdatedInfo();
+    }
+
+    private void revertSavedUserFields(){
+        if(origUserBeforeEdit != null){
+            return;
+        }
+
+        try {
+            currentUser = (User) origUserBeforeEdit.clone();
+        } catch (CloneNotSupportedException e) {
+            e.printStackTrace();
+        }
     }
 
     private void saveUpdatedInfo() {
@@ -174,11 +197,7 @@ public class EditProfileActivity extends AuthenticatedActivity {
                 }
             }
             else{
-                try {
-                    currentUser = (User) origUserBeforeEdit.clone();
-                } catch (CloneNotSupportedException e) {
-                    e.printStackTrace();
-                }
+                revertSavedUserFields();
             }
 
             message(response);
@@ -299,15 +318,12 @@ public class EditProfileActivity extends AuthenticatedActivity {
             changePasswordButton.setVisibility(View.GONE);
         }
         else{
-
             changePasswordButton.setVisibility(View.VISIBLE);
         }
     }
 
     private void addValidations() {
         fieldsToValidate = new ArrayList<>(Arrays.asList(
-                new NameValidation(context, editTextFirstName),
-                new NameValidation(context, editTextLastName),
                 new CityValidation(context, editCity),
                 new BrgyValidation(context, editBrgy),
                 new PhoneNumberValidation(context, editPhone)));
