@@ -3,7 +3,7 @@ package com.example.thesis.utilities;
 import android.text.TextUtils;
 
 public class Generic{
-  
+
     public static final String USER_LOGGED_IN_TAG = "userLoggedIn";
     public static final String SELECTED_VACCINE_TAG = "selectedVaccine";
     public static final String SELECTED_SCHEDULE = "selectedSchedule";
