@@ -7,6 +7,7 @@ import java.text.ParseException;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
+import java.time.format.DateTimeParseException;
 import java.util.Calendar;
 import java.util.Date;
 
@@ -23,13 +24,18 @@ public class BirthDayValidation extends DateValidation{
     @Override
     public boolean isValid(String stringDateToCheck) {
 
-        Instant dateToCheck = Instant.parse(stringDateToCheck + context.getResources().getString(R.string.instant_zero_time));
-        int minimumAge = context.getResources().getInteger(R.integer.minimum_age_for_vaccine);
+        try {
+            Instant dateToCheck = Instant.parse(stringDateToCheck + context.getResources().getString(R.string.instant_zero_time));
+            int minimumAge = context.getResources().getInteger(R.integer.minimum_age_for_vaccine);
 
-        ZonedDateTime minimumDate = serverDate.minusYears(minimumAge);
+            ZonedDateTime minimumDate = serverDate.minusYears(minimumAge);
 
-        if(dateToCheck.isAfter(minimumDate.toInstant())){
-            errorMessage = context.getResources().getString(R.string.error_age_not_allowed, minimumAge);
+            if (dateToCheck.isAfter(minimumDate.toInstant())) {
+                errorMessage = context.getResources().getString(R.string.error_age_not_allowed, minimumAge);
+                return false;
+            }
+        }
+        catch (DateTimeParseException dateTimeParseException){
             return false;
         }
 
