@@ -1,36 +1,23 @@
 package com.example.thesis.views;
 
-import android.support.v7.widget.RecyclerView;
-import android.text.Editable;
-import android.text.TextWatcher;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
 import android.widget.EditText;
+import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.thesis.R;
-import com.example.thesis.models.ChecklistItem;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnCheckedChanged;
-import butterknife.OnClick;
+import com.example.thesis.models.Question;
 
 public class ChecklistItemViewHolder extends RecyclerView.ViewHolder {
 
-    @BindView(R.id.checklist_item_text)
-    EditText noteText;
+    public TextView noteText;
 
-    @BindView(R.id.checklist_item_checkbox)
-    CheckBox isCheckedBox;
-
-    @BindView(R.id.delete_note)
-    Button deleteNoteButton;
-
-    private TextWatcher textWatcher;
+    public CheckBox isCheckedBox;
 
     public static ChecklistItemViewHolder newInstance(ViewGroup parent) {
         View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.checklist_note_list_item, parent, false);
@@ -39,29 +26,13 @@ public class ChecklistItemViewHolder extends RecyclerView.ViewHolder {
 
     private ChecklistItemViewHolder(View itemView) {
         super(itemView);
-        ButterKnife.bind(this, itemView);
+        noteText = itemView.findViewById(R.id.checklist_item_label);
+        isCheckedBox = itemView.findViewById(R.id.checklist_item_checkbox);
     }
 
-    public void setChecklistItem(ChecklistItem item, final int position, final OnCheckListItemChangedListener onCheckListItemChangedListener) {
-        noteText.removeTextChangedListener(textWatcher);
-        noteText.setText(item.getNote());
-        textWatcher = new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-            }
+    public void setChecklistItem(Question item, final int position, final OnCheckListItemChangedListener onCheckListItemChangedListener) {
+        noteText.setText(item.getQuestion());
 
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                if (onCheckListItemChangedListener != null) {
-                    onCheckListItemChangedListener.onNoteChanged(position, s.toString());
-                }
-            }
-        };
-        noteText.addTextChangedListener(textWatcher);
         isCheckedBox.setOnCheckedChangeListener(null);
         isCheckedBox.setChecked(item.isChecked());
         isCheckedBox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
@@ -72,19 +43,9 @@ public class ChecklistItemViewHolder extends RecyclerView.ViewHolder {
                 }
             }
         });
-        deleteNoteButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (onCheckListItemChangedListener != null) {
-                    onCheckListItemChangedListener.onNoteDelete(position);
-                }
-            }
-        });
     }
 
     public interface OnCheckListItemChangedListener {
-        void onNoteChanged(int position, String note);
         void onCheckChanged(int position, boolean checked);
-        void onNoteDelete(int position);
     }
 }
